Move `composite` into `framework`
