--- docs/apidocs/composite.rst
@@ -1,6 +1,0 @@
-.. _qiskit-experiments-composite:
-
-.. automodule:: qiskit_experiments.composite
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
