+++ qiskit_experiments/framework/composite/composite_analysis.py
@@ -15,7 +15,7 @@ Composite Experiment Analysis class.
 
 from qiskit.exceptions import QiskitError
 from qiskit_experiments.framework import BaseAnalysis, AnalysisResultData
-from qiskit_experiments.composite.composite_experiment_data import CompositeExperimentData
+from .composite_experiment_data import CompositeExperimentData
 
 
 class CompositeAnalysis(BaseAnalysis):
