+++ qiskit_experiments/test/t2ramsey_backend.py
@@ -22,7 +22,7 @@ from qiskit.providers.options import Options
 from qiskit.providers.models import QasmBackendConfiguration
 from qiskit.result import Result
 from qiskit.test import QiskitTestCase
-from qiskit_experiments.composite import ParallelExperiment
+from qiskit_experiments.framework import ParallelExperiment
 from qiskit_experiments.library.characterization import T2Ramsey
 from qiskit_experiments.test.utils import FakeJob
