+++ test/calibration/experiments/test_rabi.py
@@ -23,13 +23,12 @@ from qiskit.test import QiskitTestCase
 from qiskit.qobj.utils import MeasLevel
 import qiskit.pulse as pulse
 
-from qiskit_experiments.framework import ExperimentData
+from qiskit_experiments.framework import ExperimentData, ParallelExperiment
 from qiskit_experiments.library import Rabi, EFRabi
 
 from qiskit_experiments.library.calibration.analysis.oscillation_analysis import OscillationAnalysis
 from qiskit_experiments.data_processing.data_processor import DataProcessor
 from qiskit_experiments.data_processing.nodes import Probability
-from qiskit_experiments import ParallelExperiment
 from qiskit_experiments.test.mock_iq_backend import MockIQBackend
