+++ test/test_tomography.py
@@ -23,7 +23,7 @@ from qiskit.test import QiskitTestCase
 from qiskit import QuantumCircuit
 import qiskit.quantum_info as qi
 from qiskit.providers.aer import AerSimulator
-from qiskit_experiments import BatchExperiment, ParallelExperiment
+from qiskit_experiments.framework import BatchExperiment, ParallelExperiment
 from qiskit_experiments.library import StateTomography, ProcessTomography
